requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.36
alembic>=1.13.1
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
python-dotenv>=1.0.1
pillow>=11.0.0
aiofiles>=23.2.1
httpx>=0.27.0
firebase-admin>=6.5.0
pydantic[email]>=2.9.0